requirements.txt
fastapi
uvicorn
mangum
boto3
langchain-aws
crewai>=1.10.0
crewai-tools>=1.10.0
langchain
langchain-community
python-dotenv
langchain-text-splitters
beautifulsoup4
requests
pydantic
pytest
numpy
scikit-learn
PyJWT
python-multipart
opensearch-py
circuitbreaker
python-json-logger
